move lru cache to utils
